[ECP-9072] Refactor Button classes and create a new one for PayPal
