+++ Block/ApplePay/Shortcut/Button.php
@@ -22,44 +22,17 @@ use Magento\Catalog\Block\ShortcutInterface;
 use Magento\Checkout\Model\DefaultConfigProvider;
 use Magento\Customer\Model\Session as CustomerSession;
 use Magento\Framework\App\Config\ScopeConfigInterface;
-use Magento\Framework\Exception\LocalizedException;
 use Magento\Framework\UrlInterface;
 use Magento\Framework\View\Element\Template\Context;
 use Magento\Payment\Model\MethodInterface;
-use Magento\Framework\Exception\NoSuchEntityException;
 use Magento\Store\Model\StoreManagerInterface;
 
 class Button extends AbstractButton implements ShortcutInterface
 {
     /**
-     * @var DefaultConfigProvider $defaultConfigProvider
-     */
-    private $defaultConfigProvider;
-
-    /**
-     * @var UrlInterface $url
-     */
-    private $url;
-
-    /**
-     * @var CustomerSession $customerSession
-     */
-    private $customerSession;
-
-    /**
-     * @var StoreManagerInterface $storeManager
-     */
-    private $storeManager;
-
-    /**
-     * @var ScopeConfigInterface $scopeConfig
-     */
-    private $scopeConfig;
-
-    /**
      * @var ConfigurationInterface $configuration
      */
-    private $configuration;
+    private ConfigurationInterface $configuration;
 
     /**
      * Button Constructor
@@ -73,7 +46,6 @@ class Button extends AbstractButton implements ShortcutInterface
      * @param DefaultConfigProvider $defaultConfigProvider
      * @param ScopeConfigInterface $scopeConfig
      * @param AdyenHelper $adyenHelper
-     * @param AdyenConfigHelper $adyenConfigHelper
      * @param ConfigurationInterface $configuration
      * @param array $data
      */
@@ -87,7 +59,6 @@ class Button extends AbstractButton implements ShortcutInterface
         DefaultConfigProvider $defaultConfigProvider,
         ScopeConfigInterface $scopeConfig,
         AdyenHelper $adyenHelper,
-        AdyenConfigHelper $adyenConfigHelper,
         ConfigurationInterface $configuration,
         array $data = []
     ) {
@@ -100,33 +71,11 @@ class Button extends AbstractButton implements ShortcutInterface
             $storeManagerInterface,
             $scopeConfig,
             $adyenHelper,
-            $adyenConfigHelper,
+            $defaultConfigProvider,
             $data
         );
-        $this->defaultConfigProvider = $defaultConfigProvider;
-        $this->configuration = $configuration;
-    }
 
-    /**
-     * Current Quote ID for guests
-     *
-     * @return string
-     * @throws LocalizedException
-     * @throws NoSuchEntityException
-     */
-    public function getQuoteId(): string
-    {
-        try {
-            $config = $this->defaultConfigProvider->getConfig();
-            if (!empty($config['quoteData']['entity_id'])) {
-                return $config['quoteData']['entity_id'];
-            }
-        } catch (NoSuchEntityException $e) {
-            if ($e->getMessage() !== 'No such entity with cartId = ') {
-                throw $e;
-            }
-        }
-        return '';
+        $this->configuration = $configuration;
     }
 
     /**
