+++ Block/Paypal/Shortcut/Button.php
@@ -3,7 +3,7 @@
  *
  * Adyen ExpressCheckout Module
  *
- * Copyright (c) 2020 Adyen B.V.
+ * Copyright (c) 2024 Adyen N.V.
  * This file is open source and available under the MIT license.
  * See the LICENSE file for more info.
  *
@@ -11,7 +11,7 @@
  */
 declare(strict_types=1);
 
-namespace Adyen\ExpressCheckout\Block\GooglePay\Shortcut;
+namespace Adyen\ExpressCheckout\Block\Paypal\Shortcut;
 
 use Adyen\ExpressCheckout\Block\Buttons\AbstractButton;
 use Magento\Catalog\Block\ShortcutInterface;
