+++ etc/frontend/di.xml
@@ -31,6 +31,16 @@
             <argument name="payment" xsi:type="object">AdyenPaymentHppFacade</argument>
         </arguments>
     </type>
+    <type name="Adyen\ExpressCheckout\Block\Paypal\Shortcut\Button">
+        <arguments>
+            <argument name="data" xsi:type="array">
+                <item name="template" xsi:type="string">Adyen_ExpressCheckout::paypal/shortcut.phtml</item>
+                <item name="alias" xsi:type="string">adyen.paypal.mini-cart</item>
+                <item name="button_id" xsi:type="string">adyen-paypal-mini-cart</item>
+            </argument>
+            <argument name="payment" xsi:type="object">AdyenPaymentHppFacade</argument>
+        </arguments>
+    </type>
     <type name="Magento\Customer\CustomerData\SectionPoolInterface">
         <arguments>
             <argument name="sectionSourceMap" xsi:type="array">
